Prevent the snake from going back on itself
There are two ways that could happen:

If you are going up and press left and down on the same frame, pressing
left switches your movement to left, and then the check for vertical
movement passes and you start going down in the same update.

The other way is if two updates happen in one game tick.

Fix both of these by remembering the last move. Switch the direction
variable to an enum to make it easier to test the current and previous
direction, and use a switch statement to calculate the actual changes
to the position.