Implement shared pointers for stages
This allows stages to persist even when they aren't the current stage.
We don't need to handle freeing stages because shared_ptr is ref counted.

But we're still not quite there. The game continues running even though
we aren't calling it's update method! This is because the game uses a
timer. We need to disable it when pausing, and then re-enable it after.